Customize image names during image build (#4336)
The build scripts have been modified to generate images of the format $ORGANIZATION/$PREFIX-$NAME:$TAG. The default organization is eclipse, the default tag is nightly, and the default prefix is che. You can proivde --organization:<org>, --prefix:<prefix>, or --tag:<tag> on the command line to customize the images that are generated.

There is a new build-che.sh script which builds the minimum set of images to generate a new Che server, the CLI, and unit tests for these custom image names. If you build the CLI image, we overwrite /dockerfiles/cli/version/$TAG/images file to include the customized names that have been provided before this image is built, so the registry within the Docker image is accurate.

Signed-off-by: Florent BENOIT <fbenoit@codenvy.com>