CHE-7040: Add typing delay in Navigate to file dialog (#8177)
This prevents sending request to server if the file name field was updated during the timeout. Timeout in 400 ms. was chosen because It is enough to type a new symbol when typing a file name, and on the other hand this is not a big delay between end of typing and sending a request.