CHE-12329 Rework the way of retrieving editor provider by file-type (#12375)
Fixes bug when a maven file-type (without linked editor provider) is registered and another file type for pom.xml files (with linked editor provider) is registered as well, but when the pom.xml file is opened default editor provider is returned. The fix iterates registered editor providers if the the default provider is registered for the found file type.